# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Elysia API (Governance Extension)
  version: 0.4.0
  description: |
    增量 OpenAPI（审核/风控/申诉/准入/自动串联/管理员接口）。
servers:
  - url: /api
paths:
...
  /records/{id}/comments:
    post:
      summary: 创建回复卡片
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReplyRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReplyResponse'
        '403':
          description: 仅可回复公开已发布卡片，或命中风控/准入限制
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
          nullable: true
...
    AuthorPayload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
    ReplyTarget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        moodPhrase:
          type: string
        quote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        isPublic:
          type: boolean
        publicationStatus:
          type: string
          enum: [private, pending_auto, pending_manual, pending_second_review, risk_control_24h, published, rejected, needs_changes]
        author:
          $ref: '#/components/schemas/AuthorPayload'
    ReplyContext:
      type: object
      properties:
        content:
          type: string
        parentRecordId:
          type: string
          format: uuid
        rootRecordId:
          type: string
          format: uuid
        parentTarget:
          allOf:
            - $ref: '#/components/schemas/ReplyTarget'
          nullable: true
        rootTarget:
          allOf:
            - $ref: '#/components/schemas/ReplyTarget'
          nullable: true
    RecordSummaryWithReplyContext:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        mood_phrase:
          type: string
        quote:
          type: string
          nullable: true
        extra_emotions:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        visibility_intent:
          type: string
          enum: [private, public]
        publication_status:
          type: string
          enum: [private, pending_auto, pending_manual, pending_second_review, risk_control_24h, published, rejected, needs_changes]
        is_public:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        replyContext:
          allOf:
            - $ref: '#/components/schemas/ReplyContext'
          nullable: true
    CommentMeta:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        parentRecordId:
          type: string
          format: uuid
        rootRecordId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    CreateReplyRequest:
      type: object
      required: [content, moodPhrase]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 300
        moodPhrase:
          type: string
          minLength: 1
          maxLength: 140
        quote:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        extraEmotions:
          type: array
          maxItems: 8
          items:
            type: string
            maxLength: 32
        isPublic:
          type: boolean
          description: 默认 true；设为 false 时保存为私密回复卡片
    CreateReplyResponse:
      type: object
      properties:
        comment:
          $ref: '#/components/schemas/CommentMeta'
        record:
          $ref: '#/components/schemas/RecordSummaryWithReplyContext'
        publishStatus:
          type: object
          properties:
            status:
              type: string
              enum: [private, pending_auto, pending_manual, pending_second_review, risk_control_24h, published, rejected, needs_changes]
            label:
              type: string
...